requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Text processing
tiktoken>=0.5.0
nltk>=3.8.0

# Utilities
requests>=2.31.0
tenacity>=8.2.0  # For retry logic

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Optional: Local LLM support
# transformers>=4.36.0
# torch>=2.1.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
